refactor(mobile); reapply navigation changes
