+++ apps/mobile/src/features/Dashboard/DashboardNavigator.tsx
@@ -7,10 +7,8 @@ import {NetworkTag} from '~/features/Settings/useCases/changeAppSettings/ChangeN
 import {useGovernanceManagerMaker} from '~/features/Staking/Governance/common/helpers'
 import {useSelectedWallet} from '~/features/WalletManager/hooks/useSelectedWallet'
 import {useStrings} from '~/kernel/i18n/useStrings'
-import {
-  DashboardRoutes,
-  defaultStackNavigationOptions,
-} from '~/kernel/navigation/navigation'
+import {DashboardRoutes} from '~/kernel/navigation/types'
+import {defaultStackNavigationOptions} from '~/kernel/navigation/common/helpers'
 import {StakingCenter} from '../Staking/StakingCenter'
 import {Dashboard} from './Dashboard'
 import {FailedTxScreen} from './ShowFailedTxScreen/FailedTxScreen'
