+++ apps/mobile/src/features/Send/common/navigation.ts
@@ -1,10 +1,7 @@
 import {useNavigation} from '@react-navigation/native'
 import {useRef} from 'react'
 
-import {
-  AppRouteNavigation,
-  TxHistoryRouteNavigation,
-} from '~/kernel/navigation/navigation'
+import {AppRouteNavigation, TxHistoryRouteNavigation} from '~/kernel/navigation/types'
 
 export const useNavigateTo = () => {
   const navigation = useNavigation<
