+++ apps/mobile/src/features/Staking/Governance/common/navigation.ts
@@ -5,7 +5,7 @@ import {
 } from '@react-navigation/stack'
 import {useRef} from 'react'
 
-import {WalletStackRoutes} from '~/kernel/navigation/navigation'
+import {WalletStackRoutes} from '~/kernel/navigation/types'
 
 export type Routes = {
   'staking-gov-home': undefined
